requirements.txt
Django==3.2
djangorestframework==3.12.4
Pillow==8.2.0
pytest==6.2.4
pytest-django==4.3.0
pytest-factoryboy==2.1.0
pytest-flake8==1.0.7
pytest-isort==2.0.0
pytest-mock==3.6.0
requests==2.25.1
requests-mock==1.9.3
urllib3==1.26.5
celery==4.4.7
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.6.1
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
moviepy==1.0.3
numpy==1.21.0
packaging==21.0
pluggy==0.13.1
proglog==0.1.9
py==1.10.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.13.3
schedule==1.1.0
six==1.16.0
toml==0.10.2
tqdm==4.61.2
typing-extensions==3.10.0.0
zipp==3.5.0
geckodriver-autoinstaller==0.1.0
selenium==3.141.0
soupsieve==2.2.1
